requirements.txt
fastapi
uvicorn
python-dotenv
vllm
huggingface_hub
iointel
openai
faiss-cpu
chromadb
